Actions Added v0.2

+++ src/components/misc/MaterialTableUI.js
@@ -4,9 +4,9 @@ import Container from "@material-ui/core/Container";
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import {handleBackClick} from "./utility";
+import {sendMessage} from "../messages/messagesSlice";
 
 function MaterialTableUI(props) {
-
     const {columns, data, title, user, editMethod, deleteMethod} = props;
     return (
         <Container maxWidth={"xl"}>
@@ -15,15 +15,20 @@ function MaterialTableUI(props) {
                 columns={columns}
                 data={data}
                 options={{
-                    sorting: true
+                    sorting: true,
+                    filtering: true
                 }}
                 actions={[
                     rowData => ({
                         icon: "edit",
                         tooltip: "Edit",
                         onClick: (event, rowData) => {
-                            editMethod(rowData);
-                            handleBackClick(event);
+                            if (user.id !== rowData.owner) {
+                                sendMessage("You can't edit this data", "warning");
+                            } else {
+                                editMethod(rowData);
+                                handleBackClick(event);
+                            }
                         },
                         disabled: rowData.owner !== user.id
                     }),
@@ -31,7 +36,11 @@ function MaterialTableUI(props) {
                         icon: "delete",
                         tooltip: "Delete",
                         onClick: (event, rowData) => {
-                            deleteMethod(rowData.id);
+                            if (user.id !== rowData.owner) {
+                                sendMessage("You can't delete this data", "warning");
+                            } else {
+                                deleteMethod(rowData.id);
+                            }
                         },
                         disabled: rowData.owner !== user.id
                     })
